wordlen: improve accuracy
On some systems, CHAR_BIT does not equal eight. Therefore, it is an
inaccuracy to use it to convert between bytes and bits. Remove this - do
it explicitly.

It is also technically possible for a platform to define a long as
larger than the machine's word length (some older intel processors do
this for long long, to name on example). However, by definition, the
size of a raw pointer must be the machine's word length. Use sizeof(void*)
instead to detect word length.